move sent buffer code
and session error is just the error from Apple
