+++ session.go
@@ -1,21 +1,10 @@
 package apns
 
 import (
-	"container/list"
 	"errors"
 	"io"
 	"sync"
 )
-
-// NotificationResult associates an error from Apple to a notification.
-type NotificationResult struct {
-	Notif Notification
-	Err   Error
-}
-
-func (s NotificationResult) Error() string {
-	return s.Err.Error()
-}
 
 // Session to Apple's Push Notification server.
 type Session interface {
@@ -24,29 +13,6 @@ type Session interface {
 	RequeueableNotifications() []Notification
 	Disconnect()
 	Disconnected() bool
-}
-
-type buffer struct {
-	size int
-	m    sync.Mutex
-	*list.List
-}
-
-func newBuffer(size int) *buffer {
-	return &buffer{size, sync.Mutex{}, list.New()}
-}
-
-func (b *buffer) Add(v interface{}) *list.Element {
-	b.m.Lock()
-	defer b.m.Unlock()
-
-	e := b.PushBack(v)
-
-	if b.Len() > b.size {
-		b.Remove(b.Front())
-	}
-
-	return e
 }
 
 type sessionState int
@@ -66,7 +32,7 @@ type session struct {
 	st  sessionState
 	stm sync.Mutex
 
-	err NotificationResult
+	err Error
 }
 
 // NewSession creates a new session.
@@ -130,10 +96,10 @@ func (s *session) Send(n Notification) error {
 	s.b.Add(n)
 
 	// Send synchronously
-	return s.send(b)
+	return s.write(b)
 }
 
-func (s *session) send(b []byte) error {
+func (s *session) write(b []byte) error {
 	s.connm.Lock()
 	defer s.connm.Unlock()
 
@@ -149,36 +115,6 @@ func (s *session) send(b []byte) error {
 // Disconnect from gateway.
 func (s *session) Disconnect() {
 	s.transitionState(sessionStateDisconnected)
-}
-
-// RequeueableNotifications returns good notifications sent after an error.
-func (s *session) RequeueableNotifications() []Notification {
-	notifs := []Notification{}
-
-	// If still connected, return nothing
-	if s.st != sessionStateDisconnected {
-		return notifs
-	}
-
-	// Walk back to last known good notification and return the slice
-	var e *list.Element
-	for e = s.b.Front(); e != nil; e = e.Next() {
-		if n, ok := e.Value.(Notification); ok && n.Identifier == s.err.Notif.Identifier {
-			break
-		}
-	}
-
-	// Start right after errored ID and get the rest of the list
-	for e = e.Next(); e != nil; e = e.Next() {
-		n, ok := e.Value.(Notification)
-		if !ok {
-			continue
-		}
-
-		notifs = append(notifs, n)
-	}
-
-	return notifs
 }
 
 func (s *session) transitionState(st sessionState) {
@@ -200,15 +136,5 @@ func (s *session) readErrors() {
 
 	s.Disconnect()
 
-	e := NewError(p)
-
-	for cursor := s.b.Back(); cursor != nil; cursor = cursor.Prev() {
-		// Get serialized notification
-		n, _ := cursor.Value.(Notification)
-
-		// If the notification, move cursor after the trouble notification
-		if n.Identifier == e.Identifier {
-			s.err = NotificationResult{n, e}
-		}
-	}
+	s.err = NewError(p)
 }
